requirements.txt
streamlit
pandas
gspread
oauth2client
google-genai
requests